+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Bug fixes:
 * `file.path` with `NULL` as one of its arguments gives correct result (empty character vector).
 * `format.POSIXlt` uses the same time zone database as rest of the system. #29
 * `dev.control(displaylist = 'inhibit')` caused `ClassCastException`
+* `download.file` follows redirects.
 
 # 1.0 RC 9